+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes are listed by minor version, from newest to oldest. Under each minor ver
 
 * Frame.getElement() --> Frame.get()
 * Frame.addElement() --> Frame.add()
+* Width and height moved into options object
 
 *Enhanced:*